+++ Gui/TabWidget.cpp
@@ -1986,6 +1986,10 @@ TabWidget::activeIndex() const
 void
 TabWidget::setObjectName_mt_safe(const QString & str)
 {
+    std::string strs = str.toStdString();
+    if ( NATRON_PYTHON_NAMESPACE::isKeyword(strs) ) {
+        throw std::runtime_error(strs + " is a Python keyword");
+    }
     std::string oldName = objectName_mt_safe().toStdString();
     {
         QMutexLocker l(&_imp->tabWidgetStateMutex);
@@ -2004,7 +2008,7 @@ TabWidget::setObjectName_mt_safe(const QString & str)
         ss << "if hasattr(" << appID << ", '"  << oldName << "'):\n";
         ss << "    del " << appID << "." << oldName << "\n";
     }
-    ss << appID << "." << str.toStdString() << " = " << appID << ".getTabWidget('" << str.toStdString() << "')\n";
+    ss << appID << "." << strs << " = " << appID << ".getTabWidget('" << strs << "')\n";
 
     std::string script = ss.str();
     std::string err;
@@ -2103,6 +2107,11 @@ TabWidget::onTabScriptNameChanged(PanelWidget* tab,
     }
 
     std::string paneName = objectName_mt_safe().toStdString();
+    if ( NATRON_PYTHON_NAMESPACE::isKeyword(newName) ) {
+        Dialogs::errorDialog(paneName, newName + " is a Python keyword");
+
+        return;
+    }
     std::string appID = _imp->gui->getApp()->getAppIDString();
     std::stringstream ss;
     ss << "if hasattr(" << appID << "." << paneName << ",\"" << oldName << "\"):\n";
@@ -2128,6 +2137,9 @@ TabWidgetPrivate::declareTabToPython(PanelWidget* widget,
 
     if (!isViewer && !isPanel) {
         return;
+    }
+    if ( NATRON_PYTHON_NAMESPACE::isKeyword(tabName) ) {
+        throw std::runtime_error(tabName + " is a Python keyword");
     }
 
     std::string paneName = _publicInterface->objectName_mt_safe().toStdString();